Create a sample API Rate Limiter
This API Rate Limiter is used at the method level to limit the rate of
requests. Use the annotation <code>Limiting</code> at the method level
to specify the number of requests to be processed in the a particular
time window.
See <code>com.quolum.limit.RequestProcessorDriver.java</code> for the
example usage.